Use explicit config file path in end to end tests
The trick to copy the xml configuration from the resources to the configuration file then refreshing all sections with `ConfigurationManager.RefreshSection` doesn’t work when running on mono.

Explicitly specifying the config file path solves this issue and also makes the EndToEndTest class ready for .NET Core which doesn’t have AppDomain.SetupInformation.

The `SpecifyingTheConfigFilePath` test is adapted to use the default config file path instead and renamed to `UsingDefaultConfigFilePath`.